AddTwo working

+++ includes/stack.h
@@ -2,7 +2,7 @@
 #include <variant>
 #define float32_t float
 #define float64_t double
-typedef std::variant<int, long long, float32_t, float64_t> Variable;
+typedef std::variant<int32_t, int64_t, float32_t, float64_t> Variable;
 
 class Stack {
